move log to outter package
