+++ src/features/ingestion/log.py
@@ -9,7 +9,7 @@ from loguru import logger
 
 from src.lib.result import Result, catch
 
-from ._xml import ParsingError
+from .parser._xml import ParsingError
 
 __all__ = ["log_parsing"]
