+++ src/features/ingestion/parser/research_area.py
@@ -3,7 +3,7 @@ import eliot
 from src.features.ingestion.schema import ResearchArea
 
 from . import _xml as xml
-from ._log import log_parsing
+from ..log import log_parsing
 
 __all__ = ["research_area"]
